Move private contents to main files and use @private.
