+++ assets/js/googlesitekit/api/index.js
@@ -17,13 +17,123 @@
  */
 
 /**
+ * External dependencies
+ */
+import invariant from 'invariant';
+import md5 from 'md5';
+
+/**
+ * WordPress dependencies
+ */
+import apiFetch from '@wordpress/api-fetch';
+import { addQueryArgs } from '@wordpress/url';
+
+/**
  * Internal dependencies
  */
-import { deleteItem, getKeys } from './cache';
-import { createCacheKey, siteKitRequest } from './index.private';
+import {
+	deleteItem,
+	getItem,
+	getKeys,
+	setItem,
+} from 'assets/js/googlesitekit/api/cache';
+import { sortObjectProperties } from 'assets/js/util';
 
 // Caching is enabled by default.
 let cachingEnabled = true;
+
+const KEY_SEPARATOR = '::';
+
+/**
+ * Create a cache key for a set of type/identifier/datapoint values.
+ *
+ * @private
+ * @param {string} type        The data to access. One of 'core' or 'modules'.
+ * @param {string} identifier  The data identifier, eg. a module slug like `'search-console'`.
+ * @param {string} datapoint   The endpoint to request data from.
+ * @param {Object} queryParams Query params to send with the request.
+ * @return {string} The cache key to use for this set of values.
+ */
+export const createCacheKey = ( type, identifier, datapoint, queryParams = {} ) => {
+	const keySections = [ type, identifier, datapoint ].filter( ( keySection ) => {
+		return !! keySection && keySection.length;
+	} );
+
+	if (
+		keySections.length === 3 &&
+		( !! queryParams ) && ( queryParams.constructor === Object ) &&
+		Object.keys( queryParams ).length
+	) {
+		keySections.push(
+			md5( JSON.stringify( sortObjectProperties( queryParams ) ) )
+		);
+	}
+
+	return keySections.join( KEY_SEPARATOR );
+};
+
+/**
+ * Make a request to a WP REST API Site Kit endpoint.
+ *
+ * @private
+ * @param {string}  type                The data to access. One of 'core' or 'modules'.
+ * @param {string}  identifier          The data identifier, eg. a module slug like `'search-console'`.
+ * @param {string}  datapoint           The endpoint to request data from.
+ * @param {Object}  options             Options to pass to the request
+ * @param {number}  options.cacheTTL    The oldest cache data to use, in seconds.
+ * @param {Object}  options.bodyParams  Request body data to send. (Eg. used for `POST`/`PUT` request variables.)
+ * @param {number}  options.method      HTTP method to use for this request.
+ * @param {Object}  options.queryParams Query params to send with the request.
+ * @param {boolean} options.useCache    Enable or disable caching for this request only. (Caching is only used for `GET` requests.)
+ * @return {Promise} Response of HTTP request.
+ */
+export const siteKitRequest = async ( type, identifier, datapoint, {
+	bodyParams,
+	cacheTTL = 3600,
+	method = 'GET',
+	queryParams,
+	useCache = undefined,
+} = {} ) => {
+	invariant( type, '`type` argument for requests is required.' );
+	invariant( identifier, '`identifier` argument for requests is required.' );
+	invariant( datapoint, '`datapoint` argument for requests is required.' );
+
+	// Don't check for a `false`-y `useCache` value to ensure we don't fallback
+	// to the `usingCache()` behaviour when caching is manually disabled on a
+	// per-request basis.
+	const useCacheForRequest = method === 'GET' && ( useCache !== undefined ? useCache : usingCache() );
+	const cacheKey = createCacheKey( type, identifier, datapoint, queryParams );
+
+	if ( useCacheForRequest ) {
+		const { cacheHit, value } = await getItem( cacheKey, cacheTTL );
+
+		if ( cacheHit ) {
+			return value;
+		}
+	}
+
+	// Make an API request to retrieve the results.
+	try {
+		const response = await apiFetch( {
+			data: bodyParams,
+			method,
+			path: addQueryArgs(
+				`/google-site-kit/v1/${ type }/${ identifier }/data/${ datapoint }`,
+				queryParams
+			),
+		} );
+
+		if ( useCacheForRequest ) {
+			await setItem( cacheKey, response );
+		}
+
+		return response;
+	} catch ( error ) {
+		global.console.error( 'Google Site Kit API Error', error );
+
+		throw error;
+	}
+};
 
 /**
  * Get Google Site Kit data.
@@ -41,7 +151,6 @@ let cachingEnabled = true;
  * @param {Object}  options          Extra options for this request.
  * @param {number}  options.cacheTTL The oldest cache data to use, in seconds.
  * @param {boolean} options.useCache Enable or disable caching for this request only.
- *
  * @return {Promise} A promise for the `fetch` request.
  */
 export const get = async (
@@ -75,7 +184,6 @@ export const get = async (
  * @param {Object}  options             Extra options for this request.
  * @param {number}  options.method      HTTP method to use for this request.
  * @param {boolean} options.queryParams Query params to send with the request.
- *
  * @return {Promise} A promise for the `fetch` request.
  */
 export const set = async (
@@ -106,7 +214,6 @@ export const set = async (
  * but if caching is turned off it cannot be turned on for a specific request.
  *
  * @param {boolean} shouldUseCache Set to `true` to use this cache across requests; set to `false` to disable caching.
- *
  * @return {boolean} The new caching state (`true` for on, `false` for off).
  */
 export const setUsingCache = ( shouldUseCache ) => {
@@ -137,7 +244,6 @@ export const usingCache = () => {
  * @param {string} type       The data type to operate on. One of 'core' or 'modules'.
  * @param {string} identifier The data identifier, eg. a module slug like `'adsense'`.
  * @param {string} datapoint  The endpoint to invalidate cache data for.
- *
  * @return {void}
  */
 export const invalidateCache = async ( type, identifier, datapoint ) => {
